refactor: rename RefreshResult to ConnectionInfo
